don’t write any rows if release hasn’t changed

+++ b/models.py
@@ -161,7 +161,7 @@ class Release(models.Model):
     title = models.CharField(verbose_name="Название", max_length=250)
     date = models.DateField(verbose_name="Дата из мира игр", unique=True)
     updated_at = models.DateTimeField(verbose_name="Дата последнего изменения", auto_now=True)
-    hash = models.IntegerField(verbose_name="Hash of ratings in this release")
+    hash = models.BigIntegerField(verbose_name="Hash of ratings in this release")
     q = models.DecimalField(max_digits=7, decimal_places=5, verbose_name="Q coefficient for release")
 
     class Meta:
